修改專案目標、資料夾目錄與代碼架構
