+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "python.analysis.autoImportCompletions": true,
+    "python.analysis.typeCheckingMode": "basic",
+    "editor.detectIndentation": false,
+    "editor.tabSize": 4
+}
